+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Module documentation.
+- Inline fragment support.
 
 ## [0.5.0.1] - 2019-09-10
 ### Added